requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
attrs==25.3.0
bitarray==3.7.1
cachetools==4.2.4
certifi==2025.8.3
charset-normalizer==3.4.2
ckzg==2.1.4
construct==2.10.68
construct-typing==0.5.6
cytoolz==1.0.1
dj-config-url==0.1.1
Django==5.1.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.14.0
hexbytes==1.3.1
httpcore==0.16.3
httpx==0.23.3
idna==3.10
jsonalias==0.1.1
msgpack==1.1.1
multidict==6.6.4
openpyxl==3.1.5
packaging==25.0
parsimonious==0.10.0
pillow==11.3.0
propcache==0.3.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
pyunormalize==16.0.0
regex==2025.9.18
reportlab==4.4.3
requests==2.32.5
rfc3986==1.5.0
rlp==4.1.0
sniffio==1.3.1
solana==0.30.2
solders==0.18.1
sqlparse==0.5.2
toolz==1.0.0
types-cachetools==4.2.10
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.5.0
web3==7.13.0
websockets==11.0.3
yarl==1.20.1
